pyproject.toml
[project]
name = "transaction-core"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-admin>=2.0.2",
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "psycopg[binary]>=3.3.2",
    "django-stubs>=5.2.8",
    "mypy>=1.19.0",
    "celery>=5.6.0",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]
